Fixed a bug where the conference tried to purge an empty directory, cleaned gradle libraries

+++ implementation/src/main/java/main/Conference.java
@@ -1393,7 +1393,7 @@ public class Conference implements UserManagement, VotingManagement, RequestMana
     }
 
     void purgeDirectory(File dir) {
-
+        if(!dir.exists()) return;
         for(File file : dir.listFiles()) {
             if(file.isDirectory()) {
                 purgeDirectory(file);
